Move distributions, seed and synth packages to a new data package
