+++ csb/src/main/scala/edu/msstate/dasi/csb/Benchmark.scala
@@ -1,7 +1,8 @@
 package edu.msstate.dasi.csb
 
-import edu.msstate.dasi.csb.distributions.DataDistributions
+import edu.msstate.dasi.csb.data.distributions.DataDistributions
 import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
+import edu.msstate.dasi.csb.data.seed.DataParser
 import edu.msstate.dasi.csb.veracity.Veracity
 import org.apache.log4j.{Level, Logger}
 import org.apache.spark.graphx.Graph
