+++ csb/src/main/scala/edu/msstate/dasi/csb/data/synth/KroFit.scala
@@ -1,4 +1,4 @@
-package edu.msstate.dasi.csb.synth
+package edu.msstate.dasi.csb.data.synth
 
 import edu.msstate.dasi.csb.Util
 import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
